+++ src/main/java/com/example/bot/_for_shelter/repository/AnswerRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
     @Query("SELECT COUNT(a) FROM Answer a WHERE a.question.id IN :questionIds AND a.viewer.id = :id")
-    int numberReplies(@Param("questionIds") List<Integer> questionIds, @Param("id") Integer id);
+    int numberReplies(@Param("questionIds") List<Long> questionIds, @Param("id") Long id);
 
 }
